explore what survey data is causing downshift
the later years of survey data are leading to lower SSB estimates (evident in bridging plots, and through simple disabling of various survey inputs). This can't be "solved" by simply extending the rec devs, and despite the overall downshift the fits to the survey are not great in 2017/2019; I suspect that once we get the age error working right and improve survey fits the SSB will downshift even more.