add in aging error explorations
still not sure if this will make the new proposed base model. via discussions with colleagues, options are to:
1) include updated ageerr matrix, and present as simple "model data update"
2) save for future updates

Note that the improvement of this matrix means we are assuming less bias in the age reads than previous models; the SD fits are still not great, though in line with what was accepted last time around.