invoke proj routine
the SPM model is giving me very different values for total and spawning biomass, with the add'l unexpected behavior that things appear to go down in the second year even with zero catches.
i will flag this for Jim when I get a chance but for now reverting to proj.